Neue Datenbankmodel erstellt und Fehler beseitigt

+++ src/main/java/com/example/kindergarten_projekt/repository/TAnwesenheitRepository.java
@@ -1,5 +1,6 @@
 package com.example.kindergarten_projekt.repository;
 
+
 import com.example.kindergarten_projekt.model.TAnwesenheit;
 import com.example.kindergarten_projekt.model.TKinder;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface TAnwesenheitRepository extends JpaRepository<TAnwesenheit, Integer> {
+public interface TAnwesenheitRepository extends JpaRepository<TAnwesenheit, Long> {
