+++ src/main/java/com/example/kindergarten_projekt/repository/TBeitraegeRepository.java
@@ -1,7 +1,8 @@
 package com.example.kindergarten_projekt.repository;
 
+
 import com.example.kindergarten_projekt.model.TBeitraege;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TBeitraegeRepository extends JpaRepository<TBeitraege, Integer> {
+public interface TBeitraegeRepository extends JpaRepository<TBeitraege, Long> {
 }
