+++ src/main/java/com/example/kindergarten_projekt/repository/TElternRepository.java
@@ -3,5 +3,5 @@ package com.example.kindergarten_projekt.repository;
 import com.example.kindergarten_projekt.model.TEltern;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TElternRepository extends JpaRepository<TEltern,Integer> {
+public interface TElternRepository extends JpaRepository<TEltern,Long> {
 }
