+++ src/main/java/com/example/kindergarten_projekt/repository/TGruppeRepository.java
@@ -3,5 +3,5 @@ package com.example.kindergarten_projekt.repository;
 import com.example.kindergarten_projekt.model.TGruppe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TGruppeRepository extends JpaRepository<TGruppe, Integer> {
+public interface TGruppeRepository extends JpaRepository<TGruppe, Long> {
 }
