+++ src/main/java/com/example/kindergarten_projekt/repository/TKinderRepository.java
@@ -3,5 +3,5 @@ package com.example.kindergarten_projekt.repository;
 import com.example.kindergarten_projekt.model.TKinder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TKinderRepository extends JpaRepository<TKinder, Integer> {
+public interface TKinderRepository extends JpaRepository<TKinder, Long> {
 }
