+++ src/main/java/com/example/kindergarten_projekt/repository/TMitarbeiterRepository.java
@@ -3,5 +3,5 @@ package com.example.kindergarten_projekt.repository;
 import com.example.kindergarten_projekt.model.TMitarbeiter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TMitarbeiterRepository extends JpaRepository<TMitarbeiter, Integer> {
+public interface TMitarbeiterRepository extends JpaRepository<TMitarbeiter, Long> {
 }
