+++ src/main/java/com/example/kindergarten_projekt/repository/TStandortRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TStandortRepository extends JpaRepository<TStandort, Integer> {
+public interface TStandortRepository extends JpaRepository<TStandort, Long> {
     List<TStandort> id(Integer id);
 
 //    List<TStandort> findByStandortName(String standortName);
