+++ src/main/java/com/example/kindergarten_projekt/repository/TZahlungsartRepository.java
@@ -4,6 +4,6 @@ import com.example.kindergarten_projekt.model.TZahlungsart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface TZahlungsartRepository extends JpaRepository<TZahlungsart, Integer> {
+public interface TZahlungsartRepository extends JpaRepository<TZahlungsart, Long> {
 
 }
